UPBGE: Fix uv and color layers conversion.
Previously a unsigned shrot was used to access the MLoop[UV/Col]
but it result for big mesh on integer overflow and the higher
loop were not converted because not acceded.

Using a unsigned int instead fix the issue.

Fix issue: #557.

+++ source/gameengine/Converter/BL_BlenderDataConversion.cpp
@@ -337,7 +337,7 @@ SCA_IInputDevice::SCA_EnumInputs ConvertKeyCode(int key_code)
 	return gReverseKeyTranslateTable[key_code];
 }
 
-static void GetUvRgba(const RAS_MeshObject::LayerList& layers, unsigned short loop,
+static void GetUvRgba(const RAS_MeshObject::LayerList& layers, unsigned int loop,
 		MT_Vector2 uvs[RAS_Texture::MaxUnits], unsigned int rgba[RAS_ITexVert::MAX_UNIT])
 {
 	// No need to initialize layers to zero as all the converted layer are all the layers needed.
